web3j for android base for web3j android 4.1.1

+++ core/src/main/java/org/web3j/tx/ChainId.java
@@ -5,16 +5,16 @@ package org.web3j.tx;
  * <a href="https://github.com/ethereum/EIPs/blob/master/EIPS/eip-155.md">EIP-155</a>.
  */
 public class ChainId {
-    public static final byte NONE = -1;
-    public static final byte MAINNET = 1;
-    public static final byte EXPANSE_MAINNET = 2;
-    public static final byte ROPSTEN = 3;
-    public static final byte RINKEBY = 4;
-    public static final byte ROOTSTOCK_MAINNET = 30;
-    public static final byte ROOTSTOCK_TESTNET = 31;
-    public static final byte KOVAN = 42;
-    public static final byte ETHEREUM_CLASSIC_MAINNET = 61;
-    public static final byte ETHEREUM_CLASSIC_TESTNET = 62;
+    public static final int NONE = -1;
+    public static final int MAINNET = 1;
+    public static final int EXPANSE_MAINNET = 2;
+    public static final int ROPSTEN = 3;
+    public static final int RINKEBY = 4;
+    public static final int ROOTSTOCK_MAINNET = 30;
+    public static final int ROOTSTOCK_TESTNET = 31;
+    public static final int KOVAN = 42;
+    public static final int ETHEREUM_CLASSIC_MAINNET = 61;
+    public static final int ETHEREUM_CLASSIC_TESTNET = 62;
     public static final int TRUE_MAINNET = 19330;
     public static final int TRUE_TESTNET = 18928;
     public static final int TRUE_LOCALNET = 18928;
