+++ core/src/main/java/org/web3j/tx/FastRawTransactionManager.java
@@ -17,7 +17,7 @@ public class FastRawTransactionManager extends RawTransactionManager {
 
     private volatile BigInteger nonce = BigInteger.valueOf(-1);
 
-    public FastRawTransactionManager(Web3j web3j, Credentials credentials, byte chainId) {
+    public FastRawTransactionManager(Web3j web3j, Credentials credentials, int chainId) {
         super(web3j, credentials, chainId);
     }
 
@@ -32,7 +32,7 @@ public class FastRawTransactionManager extends RawTransactionManager {
     }
 
     public FastRawTransactionManager(
-            Web3j web3j, Credentials credentials, byte chainId,
+            Web3j web3j, Credentials credentials, int chainId,
             TransactionReceiptProcessor transactionReceiptProcessor) {
         super(web3j, credentials, chainId, transactionReceiptProcessor);
     }
