+++ crypto/src/main/java/org/web3j/crypto/TransactionUtils.java
@@ -29,7 +29,7 @@ public class TransactionUtils {
      * @return encoded transaction hash
      */
     public static byte[] generateTransactionHash(
-            RawTransaction rawTransaction, byte chainId, Credentials credentials) {
+            RawTransaction rawTransaction, int chainId, Credentials credentials) {
         byte[] signedMessage = TransactionEncoder.signMessage(rawTransaction, chainId, credentials);
         return Hash.sha3(signedMessage);
     }
@@ -55,7 +55,7 @@ public class TransactionUtils {
      * @return transaction hash as a hex encoded string
      */
     public static String generateTransactionHashHexEncoded(
-            RawTransaction rawTransaction, byte chainId, Credentials credentials) {
+            RawTransaction rawTransaction, int chainId, Credentials credentials) {
         return Numeric.toHexString(generateTransactionHash(rawTransaction, chainId, credentials));
     }
 }
